Refactor to rely on KapuaLocator to get services
