+++ commons/src/main/java/org/eclipse/kapua/commons/configuration/AbstractKapuaConfigurableService.java
@@ -48,9 +48,8 @@ import java.util.Properties;
 
 /**
  * Configurable service definition abstract reference implementation.
- * 
- * @since 1.0
  *
+ * @since 1.0
  */
 @SuppressWarnings("serial")
 public abstract class AbstractKapuaConfigurableService extends AbstractKapuaService implements KapuaConfigurableService, Serializable {
@@ -60,7 +59,7 @@ public abstract class AbstractKapuaConfigurableService extends AbstractKapuaServ
 
     /**
      * Reads metadata for the service pid
-     * 
+     *
      * @param pid
      * @return
      * @throws IOException
@@ -76,16 +75,17 @@ public abstract class AbstractKapuaConfigurableService extends AbstractKapuaServ
 
         String metatypeXmlName = sbMetatypeXmlName.toString();
         URL metatypeXmlURL = ResourceUtils.getResource(metatypeXmlName);
-        String metatypeXml = ResourceUtils.readResource(metatypeXmlURL);
-        if (metatypeXml != null) {
+        if (metatypeXmlURL != null) {
+            String metatypeXml = ResourceUtils.readResource(metatypeXmlURL);
             metaData = XmlUtil.unmarshal(metatypeXml, KapuaTmetadata.class);
         }
+
         return metaData;
     }
 
     /**
      * Validate configuration
-     * 
+     *
      * @param pid
      * @param ocd
      * @param updatedProps
@@ -151,7 +151,7 @@ public abstract class AbstractKapuaConfigurableService extends AbstractKapuaServ
 
     /**
      * Convert the properties map to {@link Properties}
-     * 
+     *
      * @param values
      * @return
      */
@@ -165,7 +165,7 @@ public abstract class AbstractKapuaConfigurableService extends AbstractKapuaServ
 
     /**
      * Convert the {@link Properties} to a properties map
-     * 
+     *
      * @param ocd
      * @param props
      * @return
@@ -185,7 +185,7 @@ public abstract class AbstractKapuaConfigurableService extends AbstractKapuaServ
 
     /**
      * Create the service configuration entity
-     * 
+     *
      * @param em
      * @param serviceConfig
      * @return
@@ -212,7 +212,7 @@ public abstract class AbstractKapuaConfigurableService extends AbstractKapuaServ
 
     /**
      * Update the service configuration entity
-     * 
+     *
      * @param em
      * @param serviceConfig
      * @return
@@ -243,7 +243,7 @@ public abstract class AbstractKapuaConfigurableService extends AbstractKapuaServ
 
     /**
      * Constructor
-     * 
+     *
      * @param pid
      * @param domain
      * @param entityManagerFactory
@@ -266,7 +266,7 @@ public abstract class AbstractKapuaConfigurableService extends AbstractKapuaServ
 
         try {
             KapuaTmetadata metadata = readMetadata(this.pid);
-            if (metadata.getOCD() != null && metadata.getOCD().size() > 0) {
+            if (metadata != null && metadata.getOCD() != null && metadata.getOCD().size() > 0) {
                 for (KapuaTocd ocd : metadata.getOCD()) {
                     if (ocd.getId() != null && ocd.getId().equals(pid)) {
                         return ocd;
