+++ console/src/main/java/org/eclipse/kapua/app/console/shared/service/GwtDomainService.java
@@ -15,10 +15,8 @@ package org.eclipse.kapua.app.console.shared.service;
 import com.google.gwt.user.client.rpc.RemoteService;
 import com.google.gwt.user.client.rpc.RemoteServiceRelativePath;
 import org.eclipse.kapua.app.console.shared.GwtKapuaException;
-import org.eclipse.kapua.app.console.shared.model.GwtConfigComponent;
 import org.eclipse.kapua.app.console.shared.model.GwtPermission.GwtAction;
 import org.eclipse.kapua.app.console.shared.model.GwtPermission.GwtDomain;
-import org.eclipse.kapua.app.console.shared.model.GwtXSRFToken;
 
 import java.util.List;
 
@@ -46,21 +44,4 @@ public interface GwtDomainService extends RemoteService {
     public List<GwtAction> findActionsByDomainName(String domainName)
         throws GwtKapuaException;
 
-    /**
-     * Returns the configuration of a Device as the list of all the configurable components.
-     *
-     * @param scopeId
-     * @return
-     */
-    public List<GwtConfigComponent> findServiceConfigurations(String scopeId)
-            throws GwtKapuaException;
-
-    /**
-     * Updates the configuration of the provided component.
-     *
-     * @param scopeId
-     * @param configComponent
-     */
-    public void updateComponentConfiguration(GwtXSRFToken xsrfToken, String scopeId, GwtConfigComponent configComponent)
-            throws GwtKapuaException;
 }
